refactor(global): make use of uint8 in command port handling

+++ cmd/http.go
@@ -109,7 +109,7 @@ func addDecoder(router *http.ServeMux, path string, decoder decoder.Decoder) {
 func getHandler(decoder decoder.Decoder) func(http.ResponseWriter, *http.Request) {
 	return func(w http.ResponseWriter, r *http.Request) {
 		type request struct {
-			Port    int16  `json:"port" validate:"required,gt=0,lte=255"`
+			Port    uint8  `json:"port" validate:"required,gt=0,lte=255"`
 			Payload string `json:"payload" validate:"required,hexadecimal"`
 			DevEUI  string `json:"devEui" validate:"omitempty,hexadecimal,len=16"`
 		}
@@ -146,12 +146,12 @@ func getHandler(decoder decoder.Decoder) func(http.ResponseWriter, *http.Request
 			if errors.Is(err, helpers.ErrValidationFailed) {
 				warnings = []string{}
 				for _, err := range helpers.UnwrapError(err) {
-					logger.Logger.Warn("validation error", zap.Error(err), zap.String("devEui", req.DevEUI), zap.Int16("port", req.Port))
+					logger.Logger.Warn("validation error", zap.Error(err), zap.String("devEui", req.DevEUI), zap.Uint8("port", req.Port))
 					warnings = append(warnings, err.Error())
 				}
 				logger.Logger.Warn("validation for some fields failed - are you using the correct port?")
 			} else {
-				logger.Logger.Error("error while decoding payload", zap.Error(err), zap.String("devEui", req.DevEUI), zap.Int16("port", req.Port))
+				logger.Logger.Error("error while decoding payload", zap.Error(err), zap.String("devEui", req.DevEUI), zap.Uint8("port", req.Port))
 
 				setBody(w, http.StatusBadRequest, map[string]any{
 					"error": err.Error(),
@@ -161,7 +161,7 @@ func getHandler(decoder decoder.Decoder) func(http.ResponseWriter, *http.Request
 			}
 		}
 
-		logger.Logger.Info("payload decoded successfully", zap.String("devEui", req.DevEUI), zap.Int16("port", req.Port))
+		logger.Logger.Info("payload decoded successfully", zap.String("devEui", req.DevEUI), zap.Uint8("port", req.Port))
 		setBody(w, http.StatusOK, map[string]any{
 			"data":     data.Data,
 			"metadata": data.Metadata,
