+++ pkg/decoder/nomadxs/v1/decoder_test.go
@@ -14,7 +14,7 @@ import (
 func TestDecode(t *testing.T) {
 	tests := []struct {
 		payload     string
-		port        int16
+		port        uint8
 		autoPadding bool
 		expected    any
 	}{
@@ -202,7 +202,7 @@ func TestDecode(t *testing.T) {
 func TestValidationErrors(t *testing.T) {
 	tests := []struct {
 		payload  string
-		port     int16
+		port     uint8
 		expected error
 	}{
 		{
@@ -319,7 +319,7 @@ func TestPayloadTooLong(t *testing.T) {
 func TestFeatures(t *testing.T) {
 	tests := []struct {
 		payload        string
-		port           int16
+		port           uint8
 		skipValidation bool
 	}{
 		{
@@ -454,7 +454,7 @@ func TestFeatures(t *testing.T) {
 func TestMarshal(t *testing.T) {
 	tests := []struct {
 		payload  string
-		port     int16
+		port     uint8
 		expected []string
 	}{
 		{
