+++ pkg/decoder/tagxl/v1/decoder_test.go
@@ -65,7 +65,7 @@ func TestDecode(t *testing.T) {
 
 	tests := []struct {
 		payload     string
-		port        int16
+		port        uint8
 		devEui      string
 		autoPadding bool
 		expected    any
@@ -316,7 +316,7 @@ func TestDecode(t *testing.T) {
 func TestValidationErrors(t *testing.T) {
 	tests := []struct {
 		payload  string
-		port     int16
+		port     uint8
 		expected error
 	}{}
 
@@ -367,7 +367,7 @@ func TestPayloadTooLong(t *testing.T) {
 func TestFeatures(t *testing.T) {
 	tests := []struct {
 		payload        string
-		port           int16
+		port           uint8
 		skipValidation bool
 	}{
 		{
@@ -483,7 +483,7 @@ func TestFeatures(t *testing.T) {
 func TestMarshal(t *testing.T) {
 	tests := []struct {
 		payload  string
-		port     int16
+		port     uint8
 		expected []string
 	}{
 		{
